src/requirements.txt
llama-index-core
llama-index-llms-google-genai
python-dotenv
tavily-python
